pre-format

+++ src/parsing.cpp
@@ -16,13 +16,8 @@
 #include <unordered_map>
 #include <vector>
 
-static const std::array<std::string, static_cast<size_t>(grouping::_count)> grouping_lookup = {
-  "Grouped",
-  "Sampled",
-  "Delayed",
-  "Candle1Minute"
-};
-
+static const std::array<std::string, static_cast<size_t>(grouping::_count)>
+    grouping_lookup = {"Grouped", "Sampled", "Delayed", "Candle1Minute"};
 
 constexpr std::array<grouping, 256> faster_grouping_lookup = [] {
   std::array<grouping, 256> arr = {};
@@ -33,36 +28,32 @@ constexpr std::array<grouping, 256> faster_grouping_lookup = [] {
   return arr;
 }();
 
-static const std::array<std::string, static_cast<size_t>(direction::_count)> direction_lookup = {
-  "up",
-  "down",
-  "unchanged",
+static const std::array<std::string, static_cast<size_t>(direction::_count)>
+    direction_lookup = {
+        "up",
+        "down",
+        "unchanged",
 };
 
 // Convert direction enum to string for output
-std::string_view to_string(direction dir) {
-  return to_cstring(dir);
-}
+std::string_view to_string(direction dir) { return to_cstring(dir); }
 
 const std::string &to_cstring(direction dir) {
   return direction_lookup[static_cast<size_t>(dir)];
 }
 
 // Convert price_type enum to string for output
-std::string_view to_string(grouping pt) {
-  return to_cstring(pt);
-}
+std::string_view to_string(grouping pt) { return to_cstring(pt); }
 
 const std::string &to_cstring(grouping pt) {
   return grouping_lookup[static_cast<size_t>(pt)];
 }
 
-
 // String to price_type conversion helper
 grouping string_to_price_type(std::string_view key) {
   // auto it = grouping_map.find(key);
   // return (it != grouping_map.end()) ? it->second : grouping::unknown;
-  return faster_grouping_lookup[key[0]];
+  return faster_grouping_lookup[static_cast<size_t>(key[0])];
 }
 
 // Parse a comma-separated price string into a tick object
@@ -106,32 +97,29 @@ tick parse_tick(const std::string &price_string, grouping price_type) {
       (windows_ticks - WINDOWS_TICKS_TO_UNIX_EPOCH) * TICKS_PER_NANOSECOND;
 
   auto timestamp_value = std::chrono::time_point<std::chrono::system_clock,
-    std::chrono::nanoseconds>{
-    std::chrono::nanoseconds(unix_time_ns)
-  };
+                                                 std::chrono::nanoseconds>{
+      std::chrono::nanoseconds(unix_time_ns)};
 
   auto now = std::chrono::system_clock::now(); // Ensure system clock is in UTC
   auto latency_value = std::chrono::duration_cast<std::chrono::nanoseconds>(
-    now - timestamp_value);
+      now - timestamp_value);
 
   // Create tick using aggregate initialization with designated initializers
-  tick result{
-    .quote_id = std::stoi(fields[0]),
-    .bid = std::stod(fields[1]),
-    .ask = std::stod(fields[2]),
-    .daily_change = std::stod(fields[3]),
-    .dir = dir_value,
-    .tradable = fields[5] == "1",
-    .high = std::stod(fields[6]),
-    .low = std::stod(fields[7]),
-    .hash = fields[8],
-    .call_only = fields[9] == "1",
-    .mid_price = std::stod(fields[10]),
-    .timestamp = timestamp_value,
-    .field13 = std::stoi(fields[12]),
-    .group = price_type,
-    .latency = latency_value
-  };
+  tick result{.quote_id = std::stoi(fields[0]),
+              .bid = std::stod(fields[1]),
+              .ask = std::stod(fields[2]),
+              .daily_change = std::stod(fields[3]),
+              .dir = dir_value,
+              .tradable = fields[5] == "1",
+              .high = std::stod(fields[6]),
+              .low = std::stod(fields[7]),
+              .hash = fields[8],
+              .call_only = fields[9] == "1",
+              .mid_price = std::stod(fields[10]),
+              .timestamp = timestamp_value,
+              .field13 = std::stoi(fields[12]),
+              .group = price_type,
+              .latency = latency_value};
 
   return result;
 }
